+++ .env
@@ -1,1 +1,2 @@
 API_URL=http://172.23.240.1:3000/api
+API_IMG=http://localhost:3000/storage
