+++ lib/core/config/environment/environment.dart
@@ -5,4 +5,5 @@ class Environment{
     await dotenv.load(fileName: '.env');
   }
   static String apiUrl = dotenv.env['API_URL'] ?? 'the api key is not configure';
+  static String apiImg = dotenv.env['API_IMG'] ?? 'the api IMG is not configure';
 }
