--- lib/feacture/home/domain/entities/card_slide_data.dart
@@ -1,19 +1,0 @@
-class CardSlideData {
-  final String backgroundImage;
-  final double price;
-  final String profession;
-  final double qualification;
-  final String profesionalName;
-  final String profileImage;
-  final String shortDescription;
-
-  CardSlideData({
-    required this.backgroundImage,
-    required this.price,
-    required this.profession,
-    required this.qualification,
-    required this.profesionalName,
-    required this.profileImage,
-    required this.shortDescription,
-  });
-}
