--- lib/feacture/home/domain/entities/pruebas/card_slide_category.dart
@@ -1,40 +1,0 @@
-import 'package:handiman_v0/feacture/home/domain/entities/card_slide_data.dart';
-
-List<CardSlideData> categories = [
-  CardSlideData(
-      backgroundImage: 'assets/images/icon_category1.png',
-      price: 0,
-      profession: '',
-      qualification: 0,
-      profesionalName: '',
-      profileImage: 'assets/images/icon_category1.png',
-      shortDescription: 'Plumbing'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/icon_category2.png',
-      price: 0,
-      profession: '',
-      qualification: 0,
-      profesionalName: '',
-      profileImage: 'assets/images/icon_category2.png',
-      shortDescription: 'Renovation'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/icon_category3.png',
-      price: 0,
-      profession: '',
-      qualification: 0,
-      profesionalName: '',
-      profileImage: 'assets/images/icon_category3.png',
-      shortDescription: 'Painting'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/icon_category4.png',
-      price: 0,
-      profession: '',
-      qualification: 0,
-      profesionalName: '',
-      profileImage: 'assets/images/icon_category4.png',
-      shortDescription: 'Roofing'
-    ),
-];
