--- lib/feacture/home/domain/entities/pruebas/card_slide_profesional.dart
@@ -1,40 +1,0 @@
-import 'package:handiman_v0/feacture/home/domain/entities/card_slide_data.dart';
-
-List<CardSlideData> trabajadores = [
-  CardSlideData(
-      backgroundImage: 'assets/images/worker1.png',
-      price: 24.0,
-      profession: 'Oficinista Tiempo Completo',
-      qualification: 4,
-      profesionalName: '-- John Mayers --',
-      profileImage: 'assets/images/worker1.png',
-      shortDescription: 'Carpentry Services'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/worker2.png',
-      price: 24.0,
-      profession: 'Axiliar de deposito',
-      qualification: 4,
-      profesionalName: '-- Tim Keller --',
-      profileImage: 'assets/images/worker2.png',
-      shortDescription: 'Construction Engineer'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/worker3.png',
-      price: 24.0,
-      profession: 'Electricista Industrial',
-      qualification: 4,
-      profesionalName: '-- Tim Keller --',
-      profileImage: 'assets/images/worker3.png',
-      shortDescription: 'Construction Engineer'
-    ),
-    CardSlideData(
-      backgroundImage: 'assets/images/worker1.png',
-      price: 24.0,
-      profession: 'Marketing y Trader',
-      qualification: 4,
-      profesionalName: '-- Tim Keller --',
-      profileImage: 'assets/images/worker1.png',
-      shortDescription: 'Construction Engineer'
-    ),
-];
