+++ lib/features/home/domain/entities/card_worker_slide_data.dart
@@ -1,0 +1,22 @@
+class CardWorkerSlideData {
+
+  final int? id;
+  final String backgroundImage;
+  final double price;
+  final String profession;
+  final double qualification;
+  final String profesionalName;
+  final String profileImage;
+  final String shortDescription;
+
+  CardWorkerSlideData({
+    this.id,
+    required this.backgroundImage,
+    required this.price,
+    required this.profession,
+    required this.qualification,
+    required this.profesionalName,
+    required this.profileImage,
+    required this.shortDescription,
+  });
+}
