+++ lib/features/home/domain/entities/pruebas/card_slide_category.dart
@@ -1,0 +1,40 @@
+import 'package:handiman_v0/features/home/domain/entities/card_worker_slide_data.dart';
+
+List<CardWorkerSlideData> categories = [
+  CardWorkerSlideData(
+      backgroundImage: 'assets/images/icon_category1.png',
+      price: 0,
+      profession: '',
+      qualification: 0,
+      profesionalName: '',
+      profileImage: 'assets/images/icon_category1.png',
+      shortDescription: 'Plumbing'
+    ),
+    CardWorkerSlideData(
+      backgroundImage: 'assets/images/icon_category2.png',
+      price: 0,
+      profession: '',
+      qualification: 0,
+      profesionalName: '',
+      profileImage: 'assets/images/icon_category2.png',
+      shortDescription: 'Renovation'
+    ),
+    CardWorkerSlideData(
+      backgroundImage: 'assets/images/icon_category3.png',
+      price: 0,
+      profession: '',
+      qualification: 0,
+      profesionalName: '',
+      profileImage: 'assets/images/icon_category3.png',
+      shortDescription: 'Painting'
+    ),
+    CardWorkerSlideData(
+      backgroundImage: 'assets/images/icon_category4.png',
+      price: 0,
+      profession: '',
+      qualification: 0,
+      profesionalName: '',
+      profileImage: 'assets/images/icon_category4.png',
+      shortDescription: 'Roofing'
+    ),
+];
