+++ lib/features/home/domain/entities/pruebas/slide.dart
