+++ lib/main.dart
@@ -6,7 +6,6 @@ import 'package:handiman_v0/core/config/router/app_router.dart';
 import 'package:handiman_v0/core/themes/app_theme.dart';
 import 'package:handiman_v0/core/themes/theme_provider.dart';
 
-
 Future<void> main() async {
   await Environment.initEnvirionment();
   runApp(const ProviderScope(child: MyApp()));
@@ -19,11 +18,21 @@ class MyApp extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final AppTheme appTheme = ref.watch(themeNotifierProvider);
     final appRouter = ref.watch(goRouterProvider);
-    return MaterialApp.router(
-      routerConfig: appRouter,
-      debugShowCheckedModeBanner: false,
-      title: 'Handyman example',    
-      theme: appTheme.getTheme(),
-    );
+    final size = MediaQuery.of(context).size;
+    return 
+    // ScreenUtilInit(
+        // designSize:  Size(size.width, size.height),
+        // minTextAdapt: true,
+        // splitScreenMode: true,
+        // builder: (_, child) {
+        //   return 
+          MaterialApp.router(
+            routerConfig: appRouter,
+            debugShowCheckedModeBanner: false,
+            title: 'Handyman example',
+            theme: appTheme.getTheme(),
+          );
+    //     }
+    // );
   }
 }
