+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
     sdk: flutter
   flutter_dotenv: ^5.1.0
   flutter_riverpod: ^2.5.1
+  flutter_screenutil: ^5.9.3
   flutter_staggered_grid_view: ^0.7.0
   go_router: ^14.2.0
   intl: ^0.19.0
